basic video creation interface
Tested on Windows and Ubuntu. Drop an ffmpeg executable into the root
folder of the project or add it to your PATH. I am not committing the
executable to the repo to keep size down (its about 40mb * number of
platforms).

The debugVideo.mp4 is to show that it can append new framest o an
existing video. Ifyou don't specify an "appendee" video it'll create a
video from the frames you give it. This is all outlined in the code.

The following command should generate a video

node ffmpegWrapper.js "resources/images" 1 "resources/debugVideo.mp4"