added resolution param
Letterboxes inputs that aren't the same as the requested output
resoltuion.

I still have a little work on this, mainly adding the default case for
appending, which I think will have the new frames confrom to the old
video's resolution. I'll probably need the ffprobe package to do this,
since I'd have to read the video's metadata.